Extract get param data
